# Improve SortedSet DeepCopy performance