@eiriktsarpalis 2021-07-30
The performance improvements are impressive, but I'd be curious to know how much of that was contributed by the `KeyValuePairComparer` changes versus the rewriting of `DeepClone`. Would it be possible to provide results measuring each change individually?