@johnthcall 2021-07-30
@eiriktsarpalis  Below is the performance of main vs just the kvp comparer change vs both changes. If we decide to avoid the recursion in DeepCopy I'll make a separate PR for just the KVP comparer change.

|               Method | Toolchain |    N |          Mean |         Error |        StdDev | Ratio | Allocated |
|--------------------- |---------- |----- |--------------:|--------------:|--------------:|------:|----------:|
| SortedDictionaryCopy |      main |    1 |     233.25 ns |     14.931 ns |     17.195 ns |  1.00 |     392 B |
| SortedDictionaryCopy |      kvpc |    1 |     179.63 ns |      4.817 ns |      5.547 ns |  0.77 |     344 B |
| SortedDictionaryCopy |        pr |    1 |      95.33 ns |      3.673 ns |      4.229 ns |  0.41 |     168 B |
|                      |           |      |               |               |               |       |           |
| SortedDictionaryCopy |      main |   10 |   1,353.38 ns |     55.142 ns |     63.501 ns |  1.00 |   1,136 B |
| SortedDictionaryCopy |      kvpc |   10 |     600.19 ns |     30.233 ns |     34.817 ns |  0.44 |     944 B |
| SortedDictionaryCopy |        pr |   10 |     323.45 ns |     15.388 ns |     17.720 ns |  0.24 |     672 B |
|                      |           |      |               |               |               |       |           |
| SortedDictionaryCopy |      main |  100 |  16,490.14 ns |    785.823 ns |    904.955 ns |  1.00 |   7,664 B |
| SortedDictionaryCopy |      kvpc |  100 |   4,805.14 ns |     91.329 ns |     89.697 ns |  0.29 |   6,080 B |
| SortedDictionaryCopy |        pr |  100 |   2,782.22 ns |     68.882 ns |     76.562 ns |  0.17 |   5,712 B |
|                      |           |      |               |               |               |       |           |
| SortedDictionaryCopy |      main | 1000 | 212,509.48 ns | 10,768.851 ns | 11,969.549 ns |  1.00 |  72,512 B |
| SortedDictionaryCopy |      kvpc | 1000 |  48,069.80 ns |  1,324.823 ns |  1,525.668 ns |  0.23 |  56,576 B |
| SortedDictionaryCopy |        pr | 1000 |  28,287.56 ns |  1,260.076 ns |  1,451.105 ns |  0.13 |  56,112 B |