@eiriktsarpalis 2021-07-30
> If we decide to avoid the recursion in DeepCopy I'll make a separate PR for just the KVP comparer change.

Introducing one change per PR is good practice in general (easier to revert etc). The KVP change is low risk with great perf improvements in its own right, so it certainly meets the bar for .NET 6 RC1. 

@layomia @stephentoub thoughts?