@eiriktsarpalis 2021-09-20
Hi @johnthcall, looking at this PR again it seems like using recursion in this context is safe, since depth is always bounded by _2 log(n + 1) + 1_.

The existing implementation does contain a few inefficiencies (e.g. allocating two stacks and traversing twice). I'm wondering though if we could get some of the perf benefits of your approach while still avoiding recursion. For example, consider the following (untested) implementation:
```csharp
public Node DeepClone(int count)
{
#if DEBUG
    Debug.Assert(count == GetCount());
#endif
    Node newRoot = ShallowClone();

    var pendingNodes = new Stack<(Node source, Node target)>(2 * Log2(count) + 2);
    pendingNodes.Push((this, newRoot));

    while (pendingNodes.TryPop(out var next))
    {
        Node clonedNode;

        if (next.source.Left is Node left)
        {
            clonedNode = left.ShallowClone();
            next.target.Left = clonedNode;
            pendingNodes.Push((left, clonedNode));
        }

        if (next.source.Right is Node right)
        {
            clonedNode = right.ShallowClone();
            next.target.Right = clonedNode;
            pendingNodes.Push((right, clonedNode));
        }
    }

    return newRoot;
}
```
I'd be curious to see what performance of something like the above could be, compared to the recursive approach.