@johnthcall 2021-09-25
@eiriktsarpalis Your code does pass all UT and has a perf improvement and 56 byte allocation improvement however the recursive change does outperform it. I understand because of the possibly imbalanced tree that we may want to avoid recursion, let me know what you'd like to do here.

| Toolchain |    N |         Mean |        Error |       StdDev | Ratio | Allocated |
|---------- |----- |-------------:|-------------:|-------------:|------:|----------:|
|      main |    1 |     69.80 ns |     3.569 ns |     3.967 ns |  1.00 |     240 B |
|     eirik |    1 |     52.87 ns |     2.967 ns |     3.298 ns |  0.76 |     184 B |
|        pr |    1 |     32.39 ns |     2.562 ns |     2.847 ns |  0.47 |      96 B |
|           |      |              |              |              |       |           |
|      main |   10 |    302.86 ns |    13.321 ns |    14.806 ns |  1.00 |     768 B |
|     eirik |   10 |    242.44 ns |    12.855 ns |    14.804 ns |  0.81 |     712 B |
|        pr |   10 |    169.16 ns |     8.750 ns |     8.986 ns |  0.56 |     528 B |
|           |      |              |              |              |       |           |
|      main |  100 |  2,686.17 ns |   121.347 ns |   139.744 ns |  1.00 |   5,184 B |
|     eirik |  100 |  2,510.29 ns |   128.237 ns |   147.678 ns |  0.94 |   5,128 B |
|        pr |  100 |  1,548.30 ns |    68.639 ns |    76.292 ns |  0.58 |   4,848 B |
|           |      |              |              |              |       |           |
|      main | 1000 | 29,168.67 ns | 1,340.231 ns | 1,543.412 ns |  1.00 |  48,480 B |
|     eirik | 1000 | 26,338.11 ns |   989.322 ns | 1,139.305 ns |  0.91 |  48,424 B |
|        pr | 1000 | 19,071.81 ns |   745.912 ns |   858.994 ns |  0.66 |  48,048 B |