@eiriktsarpalis 2021-09-27
Thanks for running the benchmarks. Theoretically speaking RB tree depths are bounded by _O(log N)_, but I would need to spend time studying the actual implementation to see whether linear depth is possible (for example it might be possible that a maliciously crafted BinaryFormatter payload or similar could result in an imbalanced set being hydrated and triggering SO when attempting to clone).

cc @GrabYourPitchforks who might provide a security angle on using recursion in general.