@johnthcall 2021-10-06
@eiriktsarpalis I've made the change use iteration instead. I tried changing the original implementation to use a single Stack like in your sample code as below but it's performance did not improve from main in the N=1000 benchmark so I've gone forward with your code change.

```csharp
public Node DeepClone(int count)
{
#if DEBUG
    Debug.Assert(count == GetCount());
#endif
    // Breadth-first traversal to recreate nodes, preorder traversal to replicate nodes.

    var pendingNodes = new Stack<(Node source, Node target)>(2 * Log2(count) + 2);
    Node newRoot = ShallowClone();

    Node? originalCurrent = this;
    Node newCurrent = newRoot;

    while (originalCurrent != null)
    {
        pendingNodes.Push((originalCurrent, newCurrent));
        newCurrent.Left = originalCurrent.Left?.ShallowClone();
        originalCurrent = originalCurrent.Left;
        newCurrent = newCurrent.Left!;
    }

    while (pendingNodes.TryPop(out var next))
    {
        Node? originalRight = next.source.Right;
        Node? newRight = originalRight?.ShallowClone();
        next.target.Right = newRight;

        while (originalRight != null)
        {
            pendingNodes.Push((originalRight, newRight!));
            newRight!.Left = originalRight.Left?.ShallowClone();
            originalRight = originalRight.Left;
            newRight = newRight.Left;
        }
    }

    return newRoot;
}